refactor: separate visualizers and their utils
