+++ examples/visualization/visualizers/detection.py
@@ -4,8 +4,8 @@ import numpy as np
 
 from depthai_nodes.ml.messages import ImgDetectionsExtended, Lines
 
-from .colors import get_yolo_colors
-from .messages import (
+from utils.colors import get_yolo_colors
+from utils.message_parsers import (
     parse_detection_message,
     parse_line_detection_message,
     parse_yolo_kpts_message,
