+++ examples/visualization/visualizers/segmentation.py
@@ -4,8 +4,8 @@ import numpy as np
 
 from depthai_nodes.ml.messages import SegmentationMasks
 
-from .colors import get_adas_colors, get_ewasr_colors, get_selfie_colors
-from .messages import parse_fast_sam_message, parse_segmentation_message
+from utils.colors import get_adas_colors, get_ewasr_colors, get_selfie_colors
+from utils.message_parsers import parse_fast_sam_message, parse_segmentation_message
 
 
 def visualize_segmentation(
