+++ examples/visualization/visualizers/utils/message_parsers.py
@@ -6,6 +6,7 @@ from depthai_nodes.ml.messages import (
     ImgDetectionsExtended,
     Keypoints,
     Lines,
+    Map2D,
     SegmentationMasks,
 )
 
@@ -69,3 +70,9 @@ def parse_fast_sam_message(message: SegmentationMasks):
     """Parses the fast sam message and returns the masks."""
     masks = message.masks
     return masks
+
+
+def parse_map_message(message: Map2D):
+    """Parses the map message and returns the map."""
+    map = message.map
+    return map
